Update WrapError function to support formatted error messages.
This change updated the WrapError function to support formatted error messages by adding a format string and variable arguments. The function now uses fmt.Errorf with the provided format string and arguments to create a new error. If the input error is nil, the function returns a wrapped error with a message created by formatting the input arguments. The getLocation function is no longer called when the input error is nil.