Update GetFuncName to correctly handle function names.
This commit change updated the GetFuncName function in utils.go to correctly handle function names. It added a new condition to check if the function name contains a / character and, if so, extract the last part of the path to use as the function name. This allows for more accurate function name extraction, especially when working with imported functions.